requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-advanced-filters==1.3.0
django-allauth==0.44.0
django-braces @ https://github.com/galuszkak/django-braces/archive/5c6604123eaa73f53875c5dce5843a6598b7b52c.zip
django-cors-headers==3.7.0
django-filter==2.4.0
django-payline==0.11
django-rest-auth==0.9.5
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-filters==0.11.1
idna==2.10
Markdown==3.3.4
oauthlib==3.1.1
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
suds-jurko==0.6
urllib3==1.26.5
